+++ test/artifact_links_parser_test.dart
@@ -1,6 +1,6 @@
-import 'package:codemagic_app_preview/src/build.dart';
-import 'package:codemagic_app_preview/src/build_platform.dart';
-import 'package:codemagic_app_preview/src/artifact_links_parser.dart';
+import 'package:codemagic_app_preview/src/builds/artifact_links_parser.dart';
+import 'package:codemagic_app_preview/src/builds/build.dart';
+import 'package:codemagic_app_preview/src/builds/build_platform.dart';
 import 'package:codemagic_app_preview/src/environment_variable_accessor.dart';
 import 'package:test/test.dart';
